hhhhh

+++ projetpfe/server/specialité/controllerspecialité.js
@@ -1,6 +1,7 @@
 
 
 const Specialtie = require('./specialitiesSchema');
+const Medecin= require('../medecin/medecinshema');
 
 const getSpecialty = async (req, res) => {
     try {
@@ -34,13 +35,13 @@ const addSpecialty = async (req, res) => {
             return res.status(400).json({ error: "Cette spécialité existe déjà." });
         }
 
-       
-        const newSpecialty = new Specialtie({ nom, description });
+        const newSpecialty = new Specialtie({ nom, description, isSelected: req.body.isSelected || true });
         const savedSpecialty = await newSpecialty.save();
 
         res.status(201).json({
             nom: savedSpecialty.nom,
             description: savedSpecialty.description,
+            isSelected: savedSpecialty.isSelected,
         });
     } catch (error) {
         console.error(error);
@@ -50,23 +51,32 @@ const addSpecialty = async (req, res) => {
 
 
 
+
 const updateSpecialty = async (req, res) => {
-    const { nom, description } = req.body;
+    const { nom, description, isSelected } = req.body;
     try {
-        const updatedSpecialty = await Specialtie.findByIdAndUpdate(req.params.id, { nom, description }, { new: true });
+        const updatedSpecialty = await Specialtie.findByIdAndUpdate(req.params.id, { nom, description, isSelected }, { new: true });
         res.status(200).send({ message: 'Spécialité mise à jour avec succès', data: updatedSpecialty });
     } catch (err) {
         res.status(400).send({ error: `Erreur lors de la mise à jour de la spécialité: ${err.message}` });
     }
 };
 
+
 const deleteSpecialty = async (req, res) => {
     try {
+        const medecinsUsingSpecialty = await Medecin.find({ specialite: req.params.id });
+
+        if (medecinsUsingSpecialty.length > 0) {
+            return res.status(500).send({ error: "Cette spécialité est utilisée par au moins un médecin. Veuillez supprimer la référence dans la table des médecins avant de la supprimer." });
+        }
+
         const deletedSpecialty = await Specialtie.findByIdAndDelete(req.params.id);
         res.status(200).send({ message: 'Spécialité supprimée avec succès', data: deletedSpecialty });
     } catch (err) {
         res.status(400).send({ error: `Erreur lors de la suppression de la spécialité: ${err.message}` });
     }
 };
+
 
 module.exports = { getSpecialty, addSpecialty, updateSpecialty, deleteSpecialty, getSpecialtyById };
